+++ examples/complete-mysql/main.tf
@@ -50,9 +50,8 @@ module "db" {
   enabled_cloudwatch_logs_exports = ["general"]
   create_cloudwatch_log_group     = true
 
-  backup_retention_period = 0
-  skip_final_snapshot     = true
-  deletion_protection     = false
+  skip_final_snapshot = true
+  deletion_protection = false
 
   performance_insights_enabled          = true
   performance_insights_retention_period = 7
@@ -100,7 +99,7 @@ module "db_default" {
   major_engine_version = "8.0"      # DB option group
   instance_class       = "db.t4g.large"
 
-  allocated_storage = 20
+  allocated_storage = 200
 
   db_name  = "completeMysql"
   username = "complete_mysql"
@@ -133,7 +132,7 @@ module "db_disabled" {
 
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "~> 3.0"
+  version = "~> 5.0"
 
   name = local.name
   cidr = local.vpc_cidr
@@ -150,7 +149,7 @@ module "vpc" {
 
 module "security_group" {
   source  = "terraform-aws-modules/security-group/aws"
-  version = "~> 4.0"
+  version = "~> 5.0"
 
   name        = local.name
   description = "Complete MySQL example security group"