+++ main.tf
@@ -3,21 +3,11 @@ locals {
   create_db_parameter_group = var.create_db_parameter_group && var.putin_khuylo
   create_db_instance        = var.create_db_instance && var.putin_khuylo
 
-  create_random_password = local.create_db_instance && var.create_random_password
-  password               = local.create_random_password ? random_password.master_password[0].result : var.password
-
   db_subnet_group_name    = var.create_db_subnet_group ? module.db_subnet_group.db_subnet_group_id : var.db_subnet_group_name
   parameter_group_name_id = var.create_db_parameter_group ? module.db_parameter_group.db_parameter_group_id : var.parameter_group_name
 
   create_db_option_group = var.create_db_option_group && var.engine != "postgres"
   option_group           = local.create_db_option_group ? module.db_option_group.db_option_group_id : var.option_group_name
-}
-
-resource "random_password" "master_password" {
-  count = local.create_random_password ? 1 : 0
-
-  length  = var.random_password_length
-  special = false
 }
 
 module "db_subnet_group" {
@@ -84,11 +74,24 @@ module "db_instance" {
 
   db_name                             = var.db_name
   username                            = var.username
-  password                            = local.password
+  password                            = var.manage_master_user_password ? null : var.password
   port                                = var.port
   domain                              = var.domain
+  domain_auth_secret_arn              = var.domain_auth_secret_arn
+  domain_dns_ips                      = var.domain_dns_ips
+  domain_fqdn                         = var.domain_fqdn
   domain_iam_role_name                = var.domain_iam_role_name
+  domain_ou                           = var.domain_ou
   iam_database_authentication_enabled = var.iam_database_authentication_enabled
+  custom_iam_instance_profile         = var.custom_iam_instance_profile
+  manage_master_user_password         = var.manage_master_user_password
+  master_user_secret_kms_key_id       = var.master_user_secret_kms_key_id
+
+  manage_master_user_password_rotation                   = var.manage_master_user_password_rotation
+  master_user_password_rotate_immediately                = var.master_user_password_rotate_immediately
+  master_user_password_rotation_automatically_after_days = var.master_user_password_rotation_automatically_after_days
+  master_user_password_rotation_duration                 = var.master_user_password_rotation_duration
+  master_user_password_rotation_schedule_expression      = var.master_user_password_rotation_schedule_expression
 
   vpc_security_group_ids = var.vpc_security_group_ids
   db_subnet_group_name   = local.db_subnet_group_name
@@ -96,17 +99,19 @@ module "db_instance" {
   option_group_name      = var.engine != "postgres" ? local.option_group : null
   network_type           = var.network_type
 
-  availability_zone   = var.availability_zone
-  multi_az            = var.multi_az
-  iops                = var.iops
-  storage_throughput  = var.storage_throughput
-  publicly_accessible = var.publicly_accessible
-  ca_cert_identifier  = var.ca_cert_identifier
+  availability_zone    = var.availability_zone
+  multi_az             = var.multi_az
+  iops                 = var.iops
+  storage_throughput   = var.storage_throughput
+  publicly_accessible  = var.publicly_accessible
+  ca_cert_identifier   = var.ca_cert_identifier
+  dedicated_log_volume = var.dedicated_log_volume
 
   allow_major_version_upgrade = var.allow_major_version_upgrade
   auto_minor_version_upgrade  = var.auto_minor_version_upgrade
   apply_immediately           = var.apply_immediately
   maintenance_window          = var.maintenance_window
+  blue_green_update           = var.blue_green_update
 
   snapshot_identifier              = var.snapshot_identifier
   copy_tags_to_snapshot            = var.copy_tags_to_snapshot
@@ -117,20 +122,22 @@ module "db_instance" {
   performance_insights_retention_period = var.performance_insights_retention_period
   performance_insights_kms_key_id       = var.performance_insights_enabled ? var.performance_insights_kms_key_id : null
 
-  replicate_source_db             = var.replicate_source_db
-  replica_mode                    = var.replica_mode
-  backup_retention_period         = var.backup_retention_period
-  backup_window                   = var.backup_window
-  max_allocated_storage           = var.max_allocated_storage
-  monitoring_interval             = var.monitoring_interval
-  monitoring_role_arn             = var.monitoring_role_arn
-  monitoring_role_name            = var.monitoring_role_name
-  monitoring_role_use_name_prefix = var.monitoring_role_use_name_prefix
-  monitoring_role_description     = var.monitoring_role_description
-  create_monitoring_role          = var.create_monitoring_role
+  replicate_source_db                  = var.replicate_source_db
+  replica_mode                         = var.replica_mode
+  backup_retention_period              = var.backup_retention_period
+  backup_window                        = var.backup_window
+  max_allocated_storage                = var.max_allocated_storage
+  monitoring_interval                  = var.monitoring_interval
+  monitoring_role_arn                  = var.monitoring_role_arn
+  monitoring_role_name                 = var.monitoring_role_name
+  monitoring_role_use_name_prefix      = var.monitoring_role_use_name_prefix
+  monitoring_role_description          = var.monitoring_role_description
+  create_monitoring_role               = var.create_monitoring_role
+  monitoring_role_permissions_boundary = var.monitoring_role_permissions_boundary
 
-  character_set_name = var.character_set_name
-  timezone           = var.timezone
+  character_set_name       = var.character_set_name
+  nchar_character_set_name = var.nchar_character_set_name
+  timezone                 = var.timezone
 
   enabled_cloudwatch_logs_exports        = var.enabled_cloudwatch_logs_exports
   create_cloudwatch_log_group            = var.create_cloudwatch_log_group
@@ -145,5 +152,16 @@ module "db_instance" {
   restore_to_point_in_time = var.restore_to_point_in_time
   s3_import                = var.s3_import
 
-  tags = merge(var.tags, var.db_instance_tags)
+  db_instance_tags = var.db_instance_tags
+  tags             = var.tags
+}
+
+module "db_instance_role_association" {
+  source = "./modules/db_instance_role_association"
+
+  for_each = { for k, v in var.db_instance_role_associations : k => v if var.create_db_instance }
+
+  feature_name           = each.key
+  role_arn               = each.value
+  db_instance_identifier = module.db_instance.db_instance_identifier
 }