+++ CHANGELOG.md
@@ -2,6 +2,178 @@
 
 All notable changes to this project will be documented in this file.
 
+## [6.6.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.5.5...v6.6.0) (2024-05-03)
+
+
+### Features
+
+* Support `dedicated_log_volume` ([#549](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/549)) ([49f87e0](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/49f87e0d35a4cebafb1b2a123abf21d0909c4bf1))
+
+## [6.5.5](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.5.4...v6.5.5) (2024-04-22)
+
+
+### Bug Fixes
+
+* Add upgrade doc and fix variable description ([#548](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/548)) ([97f6261](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/97f6261ac90c6c271964185538d8d9e19728bd54))
+
+## [6.5.4](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.5.3...v6.5.4) (2024-03-21)
+
+
+### Bug Fixes
+
+* Restore aws to required providers ([#546](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/546)) ([7f37ff8](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/7f37ff8c160e819ae6766d6bf25f0fb8cec978f5))
+
+## [6.5.3](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.5.2...v6.5.3) (2024-03-19)
+
+
+### Bug Fixes
+
+* Separate `db_instance_tags` from merged `tags` sub-module input ([#544](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/544)) ([89a5763](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/89a5763625cc60ad1d8b9afd13cfda916e88996a))
+
+## [6.5.2](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.5.1...v6.5.2) (2024-03-07)
+
+
+### Bug Fixes
+
+* Update CI workflow versions to remove deprecated runtime warnings ([#542](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/542)) ([079605a](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/079605ad6b55d201393bb8225f142129e2fc2195))
+
+### [6.5.1](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.5.0...v6.5.1) (2024-03-04)
+
+
+### Bug Fixes
+
+* Allow managing `allocated_storage` for replicas ([#534](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/534)) ([7b02569](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/7b02569468142ee054441ccaeaad9a196da74bb6))
+
+## [6.5.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.4.0...v6.5.0) (2024-03-02)
+
+
+### Features
+
+* Self managed active directory arguments ([#541](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/541)) ([c85d3b0](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/c85d3b0e372c725f2953b180d12b917f72df360f))
+
+## [6.4.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.3.1...v6.4.0) (2024-02-02)
+
+
+### Features
+
+* Secretsmanager secret rotation for master user password ([#537](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/537)) ([93c85ef](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/93c85efddeac23ce1dbf96f7e53d8326bca71e25))
+
+### [6.3.1](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.3.0...v6.3.1) (2024-01-10)
+
+
+### Bug Fixes
+
+* Align the default value of `manage_master_user_password` in the instance sub-module with the root module ([#531](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/531)) ([8b90616](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/8b906169527162faa8fc928aa9b5edf6cfbc1f5e))
+
+## [6.3.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.2.0...v6.3.0) (2023-11-13)
+
+
+### Features
+
+* Blue/green deployment for postgres ([#517](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/517)) ([9c18851](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/9c188514304cba0c05bfd88d219e676cb518e710))
+
+## [6.2.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.1.1...v6.2.0) (2023-10-23)
+
+
+### Features
+
+* Add `db_instance_role_association` functionality ([#508](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/508)) ([ed027d7](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/ed027d7daa2a03df909479c3becc6bb621b7193f))
+
+### [6.1.1](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.1.0...v6.1.1) (2023-08-05)
+
+
+### Bug Fixes
+
+* Change `allocated_storage` type from string to number ([#507](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/507)) ([5939ddf](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/5939ddf85ef740db8896fd475cac8f1c3fae8e8f))
+
+## [6.1.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v6.0.0...v6.1.0) (2023-07-11)
+
+
+### Features
+
+* Pass through db_instance_master_user_secret_arn ([#502](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/502)) ([f0eb9f3](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/f0eb9f3e14b61478a772dc8bb5d3f83bb0da3294))
+
+## [6.0.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.9.0...v6.0.0) (2023-06-25)
+
+
+### ⚠ BREAKING CHANGES
+
+* Replace random password generation with manage master password, min AWS provider version increased to 5.0, `id` output replaced with `identifier` (#489)
+
+### Features
+
+* Replace random password generation with manage master password, min AWS provider version increased to 5.0, `id` output replaced with `identifier` ([#489](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/489)) ([c66cd73](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/c66cd738f86b9cd0742d9838635970ef7b5743e4))
+
+## [5.9.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.8.0...v5.9.0) (2023-05-03)
+
+
+### Features
+
+* Add output for SQL Server `listener_endpoint` ([#486](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/486)) ([564fae2](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/564fae2f62400e5fd149ed8c34c53d64bf96f9d7))
+
+## [5.8.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.7.0...v5.8.0) (2023-05-03)
+
+
+### Features
+
+* Allow replicas to have a different version than primary ([#491](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/491)) ([7a0542d](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/7a0542d0281b91ee72835ad430edb64dcdb26f51))
+
+## [5.7.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.6.0...v5.7.0) (2023-05-02)
+
+
+### Features
+
+* Add support for Oracle `nchar_character_set_name` ([#487](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/487)) ([c7fd6c5](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/c7fd6c53680236e5afd2562cdfad7d050175b713))
+
+## [5.6.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.5.0...v5.6.0) (2023-02-17)
+
+
+### Features
+
+* Add support for permissions boundary on enhanced monitoring role ([#483](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/483)) ([b3e0aec](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/b3e0aec5efccbffd4e69d59bf7f7688a2c1f4ecf))
+
+## [5.5.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.4.2...v5.5.0) (2023-02-17)
+
+
+### Features
+
+* Add support for permissions boundary on enhanced monitoring role ([#482](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/482)) ([0500e12](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/0500e12acf5ab9b100fbe995915461df43e94b38))
+
+### [5.4.2](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.4.1...v5.4.2) (2023-02-12)
+
+
+### Bug Fixes
+
+* Update cloudwatch log group creation conditional ([#481](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/481)) ([927b714](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/927b714cdad3c8b1f8e28c3b0fb35f962916630f))
+
+### [5.4.1](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.4.0...v5.4.1) (2023-02-08)
+
+
+### Bug Fixes
+
+* Revert `ignore_changes` for instance `latest_restorable_time` ([#479](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/479)) ([b80ec03](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/b80ec0399348c889f54d30a11fa7c4f58f320ca5))
+
+## [5.4.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.3.1...v5.4.0) (2023-02-07)
+
+
+### Features
+
+* Add support for custom iam instance profile ([#473](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/473)) ([2795d5a](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/2795d5adb78643e77a2cd842670bbd4464d12003))
+
+### [5.3.1](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.3.0...v5.3.1) (2023-02-07)
+
+
+### Bug Fixes
+
+* Ignore changes to last_restorable_time ([#476](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/476)) ([acb3d51](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/acb3d51a6bd6494260e91d0621ff962712d1c32e))
+
+## [5.3.0](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.2.3...v5.3.0) (2023-01-28)
+
+
+### Features
+
+* Support blue/green updates ([#468](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/468)) ([af8c0ec](https://github.com/terraform-aws-modules/terraform-aws-rds/commit/af8c0ec726ab84ef34a4a6cab0f1dc45a65ebdc5))
+
 ### [5.2.3](https://github.com/terraform-aws-modules/terraform-aws-rds/compare/v5.2.2...v5.2.3) (2023-01-03)
 
 
@@ -324,7 +496,7 @@ All notable changes to this project will be documented in this file
 ## [v2.19.0] - 2020-10-15
 
 - fix: Added tflint and fixes in README ([#265](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/265))
-- docs: Change accound to account ([#261](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/261))
+- docs: Change account to account ([#261](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/261))
 
 
 <a name="v2.18.0"></a>
@@ -744,7 +916,7 @@ All notable changes to this project will be documented in this file
 <a name="v1.14.0"></a>
 ## [v1.14.0] - 2018-03-14
 
-- Updated readme with conditional creatino section
+- Updated readme with conditional creation section
 - Enable db_parameter_group_name variable ([#47](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/47))
 
 
@@ -782,7 +954,7 @@ All notable changes to this project will be documented in this file
 <a name="v1.8.0"></a>
 ## [v1.8.0] - 2018-01-23
 
-- Add availability zone to instance ressource module ([#35](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/35))
+- Add availability zone to instance resource module ([#35](https://github.com/terraform-aws-modules/terraform-aws-rds/issues/35))
 
 
 <a name="v1.7.0"></a>
@@ -852,7 +1024,7 @@ All notable changes to this project will be documented in this file
 - Merge branch 'master' of https://github.com/terraform-aws-modules/terraform-aws-rds
 - Updated README.md for MySQL RDS example
 - Updated README.md for Postgres RDS example
-- fianl-snapshot_identifier should default to nothing
+- final-snapshot_identifier should default to nothing
 - Add enhanced monitoring example to README, and add current contributors link
 - Merge branch 'master' into final_snapshot_identifier
 - Updated links using upstream path
@@ -882,7 +1054,7 @@ All notable changes to this project will be documented in this file
 - Removed unused variable
 - Improved kms_key_id description
 - Cleaned duplicated attributes & variables
-- Merge branch 'encryption' of github.com:fmartingr/terraform-aws-rds into encryption
+- Merge branch 'encryption' of github.com:formatting/terraform-aws-rds into encryption
 - Added parameters to example
 - Added kms_key_id parameter
 - Using `encrypted` to set the db_instance.storage_encrypted